+++ internal/resources/resources.go
@@ -1,0 +1,14 @@
+package resources
+
+import (
+	"github.com/camptocamp/terraform-provider-freeipa/internal/provider"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+var (
+	resources []func(p *provider.Provider) resource.Resource
+)
+
+func Resources() []func(p *provider.Provider) resource.Resource {
+	return resources
+}